# docs/swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Groccery API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://groccery-api.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/cart": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Cart"],
        "summary": "Get a list of cart items",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "count": { "type": "number" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Cart" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "xAuth": [] }],
        "tags": ["Cart"],
        "summary": "Create a new cart item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product": {
                    "type": "string",
                    "default": "6370f340cf88e3ec4c1502a8"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Cart item added successfully."
                    },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Cart" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },

          "422": {
            "description": "Validaton error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/cart/{id}": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Cart"],
        "summary": "Get a cart item with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Cart ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "xAuth": [] }],
        "tags": ["Cart"],
        "summary": "Delete a cart with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Cart ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Cart deleted successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "security": [{ "xAuth": [] }],
        "tags": ["Cart"],
        "summary": "Update a cart item quantity",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Cart ID",
            "type": "string",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": { "type": "number", "default": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Cart item updated successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "xAuth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "category": { "type": "string" },
          "brand": { "type": "string" },
          "description": { "type": "string" },
          "image": { "type": "string" },
          "weight": { "type": "string" },
          "salePrice": { "type": "number" },
          "retailPrice": { "type": "number" },
          "rating": { "type": "number" },
          "reviews": { "type": "number" },
          "createdAt": { "type": "string", "format": "date-time" },
          "modifiedAt": { "type": "string", "format": "date-time" },
          "_id": { "type": "string" }
        }
      },
...
      "Cart": {
        "properties": {
          "userId": { "type": "string" },
          "product": {
            "type": "object",
            "$ref": "#/components/schemas/Product"
          },
          "quantity": { "type": "number" },
          "_id": { "type": "string" }
        }
      },
...
      "Error": {
        "properties": {
          "status": { "type": "string", "default": "error" },
          "message": { "type": "string" }
        }
      }
    }
  }
}